requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
Mako==1.2.0
MarkupSafe==2.1.1
pkg-resources==0.0.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
typing-extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
Flask-JWT-Extended==4.3.1
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-decouple==3.5
toml==0.10.2
Werkzeug==2.0.2
